+++ hyperdbg/hypertrace/header/pt/Pt.h
@@ -31,5 +31,4 @@
 //////////////////////////////////////////////////
 
 BOOLEAN
-HyperTracePtDisable(HYPERTRACE_OPERATION_PACKETS * HyperTraceOperationRequest,
-                    BOOLEAN                        ApplyFromVmxRootMode);
+PtCheck();