+++ hyperdbg/libhyperdbg/code/debugger/commands/extension-commands/lbr.cpp
@@ -43,7 +43,7 @@ CommandLbrHelp()
  * @return VOID
  */
 BOOLEAN
-CommandLbrSendRequest(HYPERTRACE_OPERATION_PACKETS * LbrRequest)
+CommandLbrSendRequest(HYPERTRACE_LBR_OPERATION_PACKETS * LbrRequest)
 {
     BOOL  Status;
     ULONG ReturnedLength;
@@ -53,7 +53,7 @@ CommandLbrSendRequest(HYPERTRACE_OPERATION_PACKETS * LbrRequest)
         //
         // Send the request over serial kernel debugger
         //
-        if (!KdSendHyperTracePacketsToDebuggee(LbrRequest, SIZEOF_HYPERTRACE_OPERATION_PACKETS))
+        if (!KdSendHyperTraceLbrPacketsToDebuggee(LbrRequest, SIZEOF_HYPERTRACE_LBR_OPERATION_PACKETS))
         {
             return FALSE;
         }
@@ -70,14 +70,14 @@ CommandLbrSendRequest(HYPERTRACE_OPERATION_PACKETS * LbrRequest)
         // Send IOCTL
         //
         Status = DeviceIoControl(
-            g_DeviceHandle,                      // Handle to device
-            IOCTL_PERFORM_HYPERTRACE_OPERATION,  // IO Control Code (IOCTL)
-            LbrRequest,                          // Input Buffer to driver.
-            SIZEOF_HYPERTRACE_OPERATION_PACKETS, // Input buffer length
-            LbrRequest,                          // Output Buffer from driver.
-            SIZEOF_HYPERTRACE_OPERATION_PACKETS, // Length of output buffer in bytes.
-            &ReturnedLength,                     // Bytes placed in buffer.
-            NULL                                 // synchronous call
+            g_DeviceHandle,                          // Handle to device
+            IOCTL_PERFORM_HYPERTRACE_LBR_OPERATION,  // IO Control Code (IOCTL)
+            LbrRequest,                              // Input Buffer to driver.
+            SIZEOF_HYPERTRACE_LBR_OPERATION_PACKETS, // Input buffer length
+            LbrRequest,                              // Output Buffer from driver.
+            SIZEOF_HYPERTRACE_LBR_OPERATION_PACKETS, // Length of output buffer in bytes.
+            &ReturnedLength,                         // Bytes placed in buffer.
+            NULL                                     // synchronous call
         );
 
         if (!Status)
@@ -106,7 +106,7 @@ CommandLbrSendRequest(HYPERTRACE_OPERATION_PACKETS * LbrRequest)
  * @return BOOLEAN
  */
 BOOLEAN
-HyperDbgPerformLbrOperation(HYPERTRACE_OPERATION_PACKETS * LbrRequest)
+HyperDbgPerformLbrOperation(HYPERTRACE_LBR_OPERATION_PACKETS * LbrRequest)
 {
     return CommandLbrSendRequest(LbrRequest);
 }
@@ -122,7 +122,7 @@ HyperDbgPerformLbrOperation(HYPERTRACE_OPERATION_PACKETS * LbrRequest)
 VOID
 CommandLbr(vector<CommandToken> CommandTokens, string Command)
 {
-    HYPERTRACE_OPERATION_PACKETS LbrRequest = {0};
+    HYPERTRACE_LBR_OPERATION_PACKETS LbrRequest = {0};
 
     if (CommandTokens.size() != 2)
     {
@@ -135,19 +135,19 @@ CommandLbr(vector<CommandToken> CommandTokens, string Command)
 
     if (CompareLowerCaseStrings(CommandTokens.at(1), "enable"))
     {
-        LbrRequest.HyperTraceOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_ENABLE;
+        LbrRequest.LbrOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_ENABLE;
     }
     else if (CompareLowerCaseStrings(CommandTokens.at(1), "disable"))
     {
-        LbrRequest.HyperTraceOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DISABLE;
+        LbrRequest.LbrOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DISABLE;
     }
     else if (CompareLowerCaseStrings(CommandTokens.at(1), "save"))
     {
-        LbrRequest.HyperTraceOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_SAVE;
+        LbrRequest.LbrOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_SAVE;
     }
     else if (CompareLowerCaseStrings(CommandTokens.at(1), "dump"))
     {
-        LbrRequest.HyperTraceOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DUMP;
+        LbrRequest.LbrOperationType = HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DUMP;
     }
     else
     {
@@ -162,19 +162,19 @@ CommandLbr(vector<CommandToken> CommandTokens, string Command)
     //
     if (CommandLbrSendRequest(&LbrRequest))
     {
-        if (LbrRequest.HyperTraceOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_ENABLE)
+        if (LbrRequest.LbrOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_ENABLE)
         {
             ShowMessages("LBR enabled successfully\n");
         }
-        else if (LbrRequest.HyperTraceOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DISABLE)
+        else if (LbrRequest.LbrOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DISABLE)
         {
             ShowMessages("LBR disabled successfully\n");
         }
-        else if (LbrRequest.HyperTraceOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_SAVE)
+        else if (LbrRequest.LbrOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_SAVE)
         {
             ShowMessages("LBR branches are saved\n");
         }
-        else if (LbrRequest.HyperTraceOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DUMP)
+        else if (LbrRequest.LbrOperationType == HYPERTRACE_LBR_OPERATION_REQUEST_TYPE_DUMP)
         {
             ShowMessages("LBR branches are shown\n");
         }