+++ hyperdbg/libhyperdbg/header/help.h
@@ -312,6 +312,9 @@ CommandSmiHelp();
 VOID
 CommandLbrHelp();
 
+VOID
+CommandPtHelp();
+
 //
 // hwdbg commands
 //